refactor: revert extraction of header from Wizard.tsx
