+++ src/components/Wizard/Presets/index.ts
@@ -1,5 +1,4 @@
 export * from './Buttons';
 export * from './Confirmation';
-export * from './Header';
 export * from './Loading';
 export * from './NFTDetails';
