+++ src/components/Wizard/Wizard.tsx
@@ -2,19 +2,40 @@ import React, { FC } from 'react';
 
 import classNames from 'classnames';
 
-import { Buttons, Confirmation, Header, HeaderProps, Loading } from './Presets';
+import { Buttons, Confirmation, Loading } from './Presets';
+
+import { InfoTooltip } from '../InfoTooltip/InfoTooltip';
 
 import './Wizard.scss';
+
+export interface HeaderProps {
+  header?: React.ReactNode;
+  headerInfo?: string;
+  subHeader?: string;
+  sectionDivider?: boolean;
+}
 
 export interface WizardProps extends HeaderProps {
   children: React.ReactNode;
   className?: string;
 }
 
-const Container: FC<WizardProps> = ({ children, className, header, headerInfo, subHeader, sectionDivider }) => (
+const Header: FC<HeaderProps> = ({ children, header, headerInfo, subHeader, sectionDivider = true }) => (
+  <div className={classNames('zui-wizard-header')}>
+    <div className="zui-wizard-header-container">
+      <h1>{header}</h1>
+      {headerInfo && <InfoTooltip content={headerInfo} />}
+    </div>
+    {subHeader && <h2>{subHeader}</h2>}
+    {children}
+    {sectionDivider && <hr />}
+  </div>
+);
+
+const Container: FC<WizardProps> = ({ children, className, ...headerProps }) => (
   <div className={classNames('zui-wizard', className)}>
     {/* Header */}
-    {header && <Header header={header} headerInfo={headerInfo} subHeader={subHeader} sectionDivider={sectionDivider} />}
+    {headerProps.header && <Header {...headerProps} />}
 
     {/* Wizard Body */}
     {children}
